Seperate Chart Previewer + XMOD Support (#79)
This implements the SM5 effects into playback and editing at a functional level.

Ideally this should be handle as a third render mod in addition to Time/Row due to the difficulty of editing the chart while the effects are in place, or have a second notefield that only handles the notes/beatlines only to preview the effects along side C-mod/X-mod rendering.

This also implements the readahead limit of 20 beats / 960 rows to prevent rendering of all later notes during preview, as this render cut-off effect is used in some charts.

Commits by @Psycast